add env to clean and bootstrap scripts

+++ scripts/clean.sh
@@ -1,20 +1,19 @@
 #!/bin/sh
 set -e
 
-if [ -z "$TEST_GREP" ]; then
-   TEST_GREP=""
-fi
+prod=$1
 
-node="node"
+echo "Clear all dependences"
 
-if [ "$TEST_DEBUG" ]; then
-   node="node --inspect --debug-brk"
-fi
 
 rm -rf node_modules
 rm -rf yarn.lock
 rm -rf package-lock.json
 rm -rf packages/*/node_modules
-rm -rf packages/*/lib
 rm -rf packages/*/yarn.lock
 rm -rf packages/*/package-lock.json
+
+if [ "$prod"x != "--production"x ]; then
+  rm -rf packages/*/lib
+  echo "Clean all built libs"
+fi
